UC04_Assistant/requirements.txt
azure-search-documents
msal
requests
openai
python-dotenv
streamlit
pandas
openpyxl
azure-ai-projects
azure-identity
azure-ai-projects
azure-identity
